+++ console2/src/components/organisms/ProcessStatusActivity/index.tsx
@@ -21,21 +21,15 @@
 import * as React from 'react';
 import { connect, Dispatch } from 'react-redux';
 import { push as pushHistory } from 'react-router-redux';
-import { Button, Divider, Header, Icon, Message } from 'semantic-ui-react';
+import { Button, Divider, Header, Icon } from 'semantic-ui-react';
 
 import { ConcordId } from '../../../api/common';
 import { canBeCancelled, hasState, ProcessEntry } from '../../../api/process';
-import {
-    ProcessElementEvent,
-    ProcessEventEntry,
-    ProcessEventType
-} from '../../../api/process/event';
-import { AnsibleEvent } from '../../../api/process/ansible';
+import { ProcessElementEvent, ProcessEventEntry } from '../../../api/process/event';
 import { FormListEntry } from '../../../api/process/form';
-import { actions, MAX_EVENT_COUNT } from '../../../state/data/processes/poll';
-import { ProcessEvents, State } from '../../../state/data/processes/poll/types';
-import { timestampDiffMs } from '../../../utils';
-import { ProcessActionList, ProcessElementList, ProcessStatusTable } from '../../molecules';
+import { actions } from '../../../state/data/processes/poll';
+import { State } from '../../../state/data/processes/poll/types';
+import { ProcessActionList, ProcessStatusTable } from '../../molecules';
 import { CancelProcessPopup, AnsibleStatsActivity } from '../../organisms';
 
 interface ExternalProps {
@@ -46,8 +40,6 @@ interface StateProps {
     loading: boolean;
     process?: ProcessEntry;
     forms: FormListEntry[];
-    elementEvents: Array<ProcessEventEntry<ProcessElementEvent>>;
-    tooMuchData?: boolean;
 }
 
 interface DispatchProps {
@@ -110,16 +102,7 @@ class ProcessStatusActivity extends React.Component<Props> {
     }
 
     render() {
-        const {
-            loading,
-            refresh,
-            startWizard,
-            instanceId,
-            process,
-            forms,
-            elementEvents,
-            tooMuchData
-        } = this.props;
+        const { loading, refresh, startWizard, instanceId, process, forms } = this.props;
 
         // TODO replace the loading icon with something more visually pleasing
         return (
@@ -133,17 +116,6 @@ class ProcessStatusActivity extends React.Component<Props> {
                     />
                     {process && process.status}
                 </Header>
-
-                {tooMuchData && (
-                    <Message warning={true}>
-                        Looks like there's a lot of data. Only the first {MAX_EVENT_COUNT} events
-                        were loaded. Click{' '}
-                        <a href="#" onClick={(ev) => this.handleForceLoadAll(ev)}>
-                            here
-                        </a>{' '}
-                        to load all events. Note: it may take a while.
-                    </Message>
-                )}
 
                 {process && (
                     <>
@@ -165,18 +137,6 @@ class ProcessStatusActivity extends React.Component<Props> {
                 )}
 
                 <AnsibleStatsActivity instanceId={instanceId} />
-
-                {process && elementEvents.length > 0 && (
-                    <>
-                        <Divider content="Flow Events" horizontal={true} />
-                        <ProcessElementList
-                            instanceId={instanceId}
-                            events={elementEvents}
-                            processStatus={process.status}
-                            tooMuchData={tooMuchData}
-                        />
-                    </>
-                )}
             </>
         );
     }
@@ -188,80 +148,10 @@ interface StateType {
     };
 }
 
-// TODO move to selectors?
-
-// assumes that "eventDate" can be sorted lexicographically
-const makeElementEvents = (
-    eventById: ProcessEvents
-): Array<ProcessEventEntry<ProcessElementEvent>> =>
-    Object.keys(eventById)
-        .map((k) => eventById[k])
-        .filter((e) => e.eventType === ProcessEventType.ELEMENT)
-        .map((e) => e as ProcessEventEntry<ProcessElementEvent>)
-        .sort((a, b) => (a.eventDate > b.eventDate ? 1 : a.eventDate < b.eventDate ? -1 : 0));
-
-export const combinePrePostEvents = (
-    events: Array<ProcessEventEntry<AnsibleEvent | ProcessElementEvent>>
-): Array<ProcessEventEntry<AnsibleEvent | ProcessElementEvent>> => {
-    function findEvent(
-        phase: string,
-        correlationId: string
-    ): ProcessEventEntry<AnsibleEvent | ProcessElementEvent> | undefined {
-        return events.find(
-            (value) => value.data.phase === phase && value.data.correlationId === correlationId
-        );
-    }
-
-    const processed = {};
-    const result = new Array<ProcessEventEntry<AnsibleEvent | ProcessElementEvent>>();
-    events.forEach((event) => {
-        const data = event.data;
-        if (!data.correlationId) {
-            result.push(event);
-            return;
-        }
-
-        if (processed[data.correlationId]) {
-            return;
-        }
-
-        processed[data.correlationId] = true;
-
-        if (data.phase === 'pre' && data.correlationId) {
-            const postEvent = findEvent('post', data.correlationId);
-            if (postEvent) {
-                const clone = { ...event };
-                clone.data = postEvent.data;
-                clone.duration = timestampDiffMs(postEvent.eventDate, event.eventDate);
-                result.push(clone);
-            } else {
-                result.push(event);
-            }
-        } else if (data.phase === 'post' && data.correlationId) {
-            const preEvent = findEvent('pre', data.correlationId);
-            if (preEvent) {
-                const clone = { ...preEvent };
-                clone.duration = timestampDiffMs(event.eventDate, preEvent.eventDate);
-                result.push(clone);
-            } else {
-                result.push(event);
-            }
-        } else {
-            result.push(event);
-        }
-    });
-
-    return result;
-};
-
 export const mapStateToProps = ({ processes: { poll } }: StateType): StateProps => ({
     loading: poll.currentRequest.running,
     process: poll.currentRequest.response ? poll.currentRequest.response.process : undefined,
-    forms: poll.forms,
-    elementEvents: combinePrePostEvents(makeElementEvents(poll.eventById)) as Array<
-        ProcessEventEntry<ProcessElementEvent>
-    >,
-    tooMuchData: poll.currentRequest.response ? poll.currentRequest.response.tooMuchData : false
+    forms: poll.forms
 });
 
 export const mapDispatchToProps = (