+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## [Unreleased]
+## [1.3.0] - 2019-01-15
 
 ### Added